+++ app/helpers/stratification_factors_helper.rb
@@ -1,0 +1,2 @@
+module StratificationFactorsHelper
+end