+++ app/models/randomization_scheme.rb
@@ -10,13 +10,14 @@ class RandomizationScheme < ActiveRecord::Base
 
   # Model Validation
   validates_presence_of :name, :user_id, :project_id
-  validates_uniqueness_of :name, scope: [:deleted, :project_id]
+  validates_uniqueness_of :name, case_sensitive: false, scope: [:deleted, :project_id]
   validates_numericality_of :randomization_goal, greater_than_or_equal_to: 0, only_integer: true
 
   # Model Relationships
   belongs_to :user
   belongs_to :project
   has_many :block_size_multipliers, -> { where deleted: false }
+  has_many :stratification_factors, -> { where deleted: false }
   has_many :treatment_arms,         -> { where deleted: false }
 
   # Model Methods