+++ app/models/stratification_factor.rb
@@ -1,4 +1,4 @@
-class TreatmentArm < ActiveRecord::Base
+class StratificationFactor < ActiveRecord::Base
 
   # Concerns
   include Deletable
@@ -8,7 +8,6 @@ class TreatmentArm < ActiveRecord::Base
   # Model Validation
   validates_presence_of :name, :user_id, :project_id, :randomization_scheme_id
   validates_uniqueness_of :name, case_sensitive: false, scope: [:deleted, :project_id, :randomization_scheme_id]
-  validates_numericality_of :allocation, greater_than_or_equal_to: 0, only_integer: true
 
   # Model Relationships
   belongs_to :user
@@ -16,5 +15,9 @@ class TreatmentArm < ActiveRecord::Base
   belongs_to :randomization_scheme
 
   # Model Methods
+
+  def options
+    [["One", 1], ["Two", 2]]
+  end
 
 end